requirements.txt
django
dataclasses= 0.7
djangorestframework